requirements.txt
numpy
scipy
matplotlib
tqdm
numba
cupy-cuda12x